Free GCHandle when unregistering Callbacks and CallResults
Avoids a GCHandle memory leak when a Callback or CallResult is registered, unregistered, discarded and then left for garbage collection without calling Dispose.

This results in potentially multiple GCHandles being allocated for a single Callback or CallResult. An alternative solution that only allocates once would be to implement a SafeGCHandle class, but this is expected to be a very rare use case for Callbacks and CallResults, so it would be a solution of questionable worth.